requirements.txt
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-llms-ollama>=0.1.1
llama-index-llms-openai>=0.1.1
llama-index-embeddings-huggingface>=0.1.0
llama-index-readers-file>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pypdf>=3.15.1